fixed some bugs. But the main bug is not fixed yet. DTOFPoint_factory produces random output. Sourc not yet identified.

+++ src/libraries/TOF/DTOFPoint_factory.h
@@ -17,9 +17,9 @@ class DTOFPoint_factory:public JFactory<DTOFPoint>{
   DTOFPoint_factory(){};
   ~DTOFPoint_factory(){};
 
-  float VELOCITY   ;
-  float HALFPADDLE ;
-  float BARWIDTH   ;
+  double VELOCITY   ;
+  double HALFPADDLE ;
+  double BARWIDTH   ;
   
   private:
   jerror_t brun(JEventLoop *loop, int eventnumber);	///< Invoked via JEventProcessor virtual method
